app_ia/requirements.txt
transformers==4.19.4
fastapi
pymilvus==2.1.0
jax
flax
Pillow==9.0.1